Revert "[artf1301890]: Fix for compilation issue"
This reverts commit f830498141a202c2305b41326d9e06e03b9c5929.

Reason for revert: <We cannot remove this as the SE HAL utils uses libese-nxp.conf file config reading which will not be present in SN220V5E chipset>

+++ snxxx/libese-teq1/nxp-ese/utils/ese_config.h
@@ -68,8 +68,7 @@ class EseConfig {
   static unsigned getUnsigned(const std::string& key);
   static unsigned getUnsigned(const std::string& key, unsigned default_value);
   static unsigned long getUnsignedLong(const std::string& key);
-  static unsigned long getUnsignedLong(const std::string& key,
-                                       unsigned long default_value);
+  static unsigned long getUnsignedLong(const std::string& key, unsigned default_value);
 
  private:
   static EseConfig& getInstance();
